+++ R/MiscFuns.R
@@ -641,6 +641,7 @@ summary.fixest = function(object, se = NULL, cluster = NULL, dof = NULL, .vcov, 
 	# agg
 	if(!missnull(agg)){
 	    object$coeftable = aggregate(object, agg, full = TRUE, from_summary = TRUE)
+	    object$is_agg = TRUE
 	}
 
 	return(object)
@@ -2557,13 +2558,16 @@ did_means = function(fml, base, treat_var, post_var, tex = FALSE, treat_dict, di
 #'
 #' etable(res_2F, res_2F_bis)
 #'
-i = function(var, f, f2, ref, drop, keep, drop2, keep2){
+i = function(var, f, f2, ref, drop, keep, drop2, keep2, ...){
     # Used to create interactions
 
     # Later: binning (bin = 1:3 // bin = list("a" = "[abc]")). Default name is bin name (eg "1:3")
 
     # gt = function(x) cat(sfill(x, 20), ": ", -(t0 - (t0<<-proc.time()))[3], "s\n", sep = "")
     # t0 = proc.time()
+
+    validate_dots(valid_args = "f_name")
+    dots = list(...)
 
     mc = match.call()
 
@@ -2572,7 +2576,7 @@ i = function(var, f, f2, ref, drop, keep, drop2, keep2){
     }
 
     # Finding if it's a call from fixest
-    FROM_FIXEST = sys.nframe() > 5 && any(sapply(tail(sys.calls(), 6), function(x) any(grepl("fixest", deparse(x)[1], fixed = TRUE))))
+    FROM_FIXEST = is_fixest_call()
 
     # General checks
     check_arg(var, f, f2, "vector")
@@ -2601,6 +2605,10 @@ i = function(var, f, f2, ref, drop, keep, drop2, keep2){
                 stop("The arguments 'var' and 'f' must be of the same length (currently ", length(var), " vs ", length(f), ").")
             }
         }
+    }
+
+    if(!is.null(dots$f_name)){
+        f_name = dots$f_name
     }
 
     f_num = is.numeric(f)
@@ -3750,218 +3758,6 @@ demean = function(X, f, slope.vars, slope.flag, data, weights,
 }
 
 
-#' Aggregates the values of DiD coefficients a la Sun and Abraham
-#'
-#' Simple tool that aggregates the value of CATT coefficients in staggered difference-in-difference setups (see details).
-#'
-#' @param x A \code{fixest} object.
-#' @param agg A character scalar describing the variable names to be aggregated, it is pattern-based. All variables that match the pattern will be aggregated. It must be of the form \code{"(root)"}, the parentheses must be there and the resulting variable name will be \code{"root"}. You can add another root with parentheses: \code{"(root1)regex(root2)"}, in which case the resulting name is \code{"root1::root2"}. To name the resulting variable differently you can pass a named vector: \code{c("name" = "pattern")} or \code{c("name" = "pattern(root2)")}. It's a bit intricate sorry, please see the examples.
-#' @param full Logical scalar, defaults to \code{FALSE}. If \code{TRUE}, then all coefficients are returned, not only the aggregated coefficients.
-#' @param use_weights Logical, default is \code{TRUE}. If the estimation was weighted, whether the aggregation should take into account the weights. Basically if the weights reflected frequency it should be \code{TRUE}.
-#' @param ... Arguments to be passed to \code{\link[fixest]{summary.fixest}}.
-#'
-#' @details
-#' This is a function helping to replicate the estimator from Sun and Abraham (2020). You first need to perform an estimation with cohort and relative periods dummies (typically using the function \code{\link[fixest]{i}}), this leads to estimators of the cohort average treatment effect on the treated (CATT). Then you can use this function to retrieve the average treatment effect on each relative period, or for any other way you wish to aggregate the CATT.
-#'
-#' Note that contrary to the SA article, here the cohort share in the sample is considered to be a perfect measure for the cohort share in the population.
-#'
-#' @return
-#' It returns a matrix representing a table of coefficients.
-#'
-#' @references
-#' Liyang Sun and Sarah Abraham, forthcoming, "Estimating Dynamic Treatment Effects in Event Studies with Heterogeneous Treatment Effects". Journal of Econometrics.
-#'
-#' @author
-#' Laurent Berge
-#'
-#' @examples
-#'
-#' #
-#' # DiD example
-#' #
-#'
-#' # first we set up the data
-#'
-#' set.seed(1)
-#' n_group = 20
-#' n_per_group = 5
-#'
-#' id_i = paste0((1:n_group), ":", rep(1:n_per_group, each = n_group))
-#' id_t = 1:10
-#'
-#' base = expand.grid(id = id_i, year = id_t)
-#' base$group = as.numeric(gsub(":.+", "", base$id))
-#'
-#' base$year_treated = base$group
-#' base$year_treated[base$group > 10] = 10000
-#' base$treat_post = (base$year >= base$year_treated) * 1
-#' base$time_to_treatment = pmax(base$year - base$year_treated, -1000)
-#' base$treated = (base$year_treated < 10000) * 1
-#'
-#' # The effect of the treatment is cohort specific and increases with time
-#' base$y_true = base$treat_post * (1 + 1 * base$time_to_treatment - 1 * base$group)
-#' base$y = base$y_true + rnorm(nrow(base))
-#'
-#'
-#' # The controls have a time_to_treatment equal to -1000
-#'
-#' # we drop the always treated
-#' base = base[base$group > 1,]
-#'
-#' # Now we perform the estimation
-#' res_naive = feols(y ~ i(treated, time_to_treatment,
-#'                         ref = -1, drop = -1000) | id + year, base)
-#'
-#' res_cohort = feols(y ~ i(time_to_treatment, f2 = group,
-#'                          drop = c(-1, -1000)) | id + year, base)
-#'
-#' coefplot(res_naive, ylim = c(-6, 8))
-#' att_true = tapply(base$y_true, base$time_to_treatment, mean)[-1]
-#' points(-9:8 + 0.15, att_true, pch = 15, col = 2)
-#'
-#' # The aggregate effect for each period
-#' agg_coef = aggregate(res_cohort, "(ti.*nt)::(-?[[:digit:]]+)")
-#' x = c(-9:-2, 0:8) + .35
-#' points(x, agg_coef[, 1], pch = 17, col = 4)
-#' ci_low = agg_coef[, 1] - 1.96 * agg_coef[, 2]
-#' ci_up = agg_coef[, 1] + 1.96 * agg_coef[, 2]
-#' segments(x0 = x, y0 = ci_low, x1 = x, y1 = ci_up, col = 4)
-#'
-#' legend("topleft", col = c(1, 2, 4), pch = c(20, 15, 17),
-#'        legend = c("Naive", "True", "Sun & Abraham"))
-#'
-#'
-#' # The ATT
-#' aggregate(res_cohort, c("ATT" = "treatment::[^-]"))
-#' mean(base[base$treat_post == 1, "y_true"])
-#'
-#' # With etable
-#' etable(res_naive, res_cohort, agg = "(ti.*nt)::(-?[[:digit:]]+):gro")
-#'
-aggregate.fixest = function(x, agg, full = FALSE, use_weights = TRUE, ...){
-    # Aggregates the value of coefficients
-
-    check_arg(x, "class(fixest) mbt")
-    check_arg(agg, "character scalar")
-    check_arg(full, "logical scalar")
-    # => later => extend it to more than one set of vars to agg
-
-    dots = list(...)
-    from_summary = isTRUE(dots$from_summary)
-
-    is_name = !is.null(names(agg))
-
-    if(!is_name && !grepl("(", agg, fixed = TRUE)){
-        stop("Argument 'agg' must be a character in which the pattern to match must be in between parentheses. So far there are no parenthesis: please have a look at the examples.")
-    }
-
-    coef = coef(x)
-    cname = names(coef)
-
-    qui = grepl(agg, cname)
-    if(!any(qui)){
-        if(from_summary){
-            # We make it silent when aggregate is used in summary
-            # => this way we can pool calls to agg even for models that don't have it
-            # ==> useful in etable eg
-            return(x$coeftable)
-        } else {
-            stop("The argument 'agg' does not match any variable.")
-        }
-    }
-
-    if(!isTRUE(x$summary)){
-        x = summary(x, ...)
-    }
-
-    cname_select = cname[qui]
-    if(is_name){
-        root = rep(names(agg), length(cname_select))
-        val = gsub(paste0(".*", agg, ".*"), "\\1", cname_select)
-    } else {
-        root = gsub(paste0(".*", agg, ".*"), "\\1", cname_select)
-        val = gsub(paste0(".*", agg, ".*"), "\\2", cname_select)
-    }
-
-    V = x$cov.scaled
-
-    mm = model.matrix(x)
-
-    name_df = unique(data.frame(root, val, stringsAsFactors = FALSE))
-
-    c_all = c()
-    se_all = c()
-    for(i in 1:nrow(name_df)){
-
-        r = name_df[i, 1]
-        v = name_df[i, 2]
-        v_names = cname_select[root == r & val == v]
-
-        if(use_weights && !is.null(x$weights)){
-            shares = colSums(x$weights * sign(mm[, v_names, drop = FALSE]))
-        } else {
-            shares = colSums(sign(mm[, v_names, drop = FALSE]))
-        }
-
-        shares = shares / sum(shares)
-
-        # The coef
-        c_value = sum(shares * coef[v_names])
-
-        # The variance
-        n = length(v_names)
-        s1 = matrix(shares, n, n)
-        s2 = matrix(shares, n, n, byrow = TRUE)
-
-        var_value = sum(s1 * s2 * V[v_names, v_names])
-        se_value = sqrt(var_value)
-
-        c_all[length(c_all) + 1] = c_value
-        se_all[length(se_all) + 1] = se_value
-    }
-
-    # th z & p values
-    zvalue <- c_all/se_all
-    if(x$method_type == "feols" || (x$method %in% "feglm" && !x$family$family %in% c("poisson", "binomial"))){
-
-        # I have renamed t.df into G
-        t.df = attr(vcov, "G")
-
-        if(!is.null(t.df)){
-            pvalue <- 2*pt(-abs(zvalue), max(t.df - 1, 1))
-        } else {
-            pvalue <- 2*pt(-abs(zvalue), max(x$nobs - x$nparams, 1))
-        }
-
-    } else {
-        pvalue <- 2*pnorm(-abs(zvalue))
-    }
-
-    res = cbind(c_all, se_all, zvalue, pvalue)
-    if(max(nchar(val)) == 0){
-        rownames(res) = name_df[[1]]
-    } else {
-        rownames(res) = apply(name_df, 1, paste, collapse = "::")
-    }
-
-    colnames(res) = c("Estimate", "Std. Error", "t value", "Pr(>|t|)")
-
-    if(full){
-        table_origin = x$coeftable
-        i_min = min(which(qui)) - 1
-        before = if(i_min > 0) table_origin[1:i_min, , drop = FALSE] else NULL
-
-        i_after = (1:nrow(table_origin)) > i_min & !qui
-        after = if(any(i_after)) table_origin[i_after, , drop = FALSE] else NULL
-
-        res = rbind(before, res, after)
-
-        attr(res, "type") = attr(table_origin, "type")
-    }
-
-    res
-}
-
 #### ................. ####
 #### Internal Funs     ####
 ####
@@ -4266,7 +4062,6 @@ prepare_matrix = function(fml, base, fake_intercept = FALSE){
     res = do.call("cbind", data_list)
 
     colnames(res) = all_var_names
-
 
     res
 }
@@ -5132,7 +4927,7 @@ clean_interact_names = function(x){
 
     x2clean = x[who2clean]
 
-    x_split = strsplit(x2clean, "(^|(?<=[^[:alnum:]\\._]))i(nteract)?\\(|__CLEAN__", perl = TRUE)
+    x_split = strsplit(x2clean, "(^|(?<=[^[:alnum:]\\._]))(i(nteract)?|sunab(_att)?)\\(|__CLEAN__", perl = TRUE)
 
     x_left = sapply(x_split, function(v) v[1])
     x_right = sapply(x_split, function(v) v[2])
@@ -6518,6 +6313,11 @@ mat_posdef_fix = function(X, tol = 1e-10){
     return(X)
 }
 
+
+is_fixest_call = function(){
+    sys.nframe() > 5 && any(sapply(tail(sys.calls(), 7), function(x) any(grepl("fixest", deparse(x)[1], fixed = TRUE))))
+}
+
 #### ................. ####
 #### Additional Methods ####
 ####
@@ -6722,6 +6522,7 @@ logLik.fixest = function(object, ...){
 #' @inheritParams nobs.fixest
 #' @inheritParams etable
 #'
+#' @param agg Logical scalar, default is \code{TRUE}. If the coefficients of the estimation have been aggregated, whether to report the aggregated coefficients. If \code{FALSE}, the raw coefficients will be returned.
 #' @param ... Not currently used.
 #'
 #' @details
@@ -6751,11 +6552,17 @@ logLik.fixest = function(object, ...){
 #' fixef(res)
 #'
 #'
-coef.fixest = coefficients.fixest = function(object, keep, drop, order, ...){
+coef.fixest = coefficients.fixest = function(object, keep, drop, order, agg = TRUE, ...){
 
     check_arg(keep, drop, order, "NULL character vector no na")
+    check_arg(agg, "logical scalar")
 
-    res = object$coefficients
+    if(isTRUE(object$is_agg) && agg){
+        res = object$coeftable[, 1]
+        names(res) = rownames(object$coeftable)
+    } else {
+        res = object$coefficients
+    }
 
     if(!missnull(keep) || !missnull(drop) || !missnull(order)){
         cnames = names(res)